+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: Add missing `export` to the `EdgeCorePluginFactory` type definition.
+
 ## 1.11.0 (2023-10-18)
 
 - added: `EdgeAccount.fetchSwapQuotes`, to return all relevant quotes, and not just the best one.